+++ essential-real-estate/public/templates/property/print/qr-image.php
@@ -8,5 +8,5 @@ if (!defined('ABSPATH')) {
  */
 ?>
 <img class="ere__print-property-qr-image"
-     src="https://chart.googleapis.com/chart?chs=100x100&cht=qr&chl=<?php echo esc_url(get_permalink($property_id)); ?>&choe=UTF-8"
+     src="https://quickchart.io/qr?text=<?php echo esc_url(get_permalink($property_id)); ?>&size=100"
      title="<?php echo esc_attr(get_the_title($property_id)); ?>"/>